Agrego patreons

+++ CODIGO/Protocol.bas
@@ -3587,7 +3587,8 @@ Private Sub HandleCharacterCreate()
         .simbolo = Reader.ReadInt8()
         .Idle = Reader.ReadBool()
         .Navegando = Reader.ReadBool()
-        .tipoUsuario = Reader.ReadInt16()
+        .tipoUsuario = Reader.ReadInt32()
+        
         
         If (.Pos.x <> 0 And .Pos.y <> 0) Then
             If MapData(.Pos.x, .Pos.y).charindex = charindex Then
